relocate accumulation to bundle
